+++ frontend/src/components/forms/Login.js
@@ -1,5 +1,6 @@
 import React, {useState} from "react"
-import {formInputChange} from "./formInputChange";
+import {formInputChange} from "./utils/formInputChange";
+import {handleLoginSubmit} from "./utils/handleLoginSubmit";
 
 export default function Login() {
     const [formData, setFormData] = useState({
@@ -7,31 +8,8 @@ export default function Login() {
         password: "",
     });
 
-    const handleLoginSubmit = (event) => {
-        // const baseUrl = process.env.REACT_APP_BASE_BACKEND_URL; // TODO uncomment after tests
-        const baseUrl = "http://127.0.0.1:8080";
-        console.log("Form submitted with data:", formData);
-        event.preventDefault();
-
-        fetch(baseUrl + "/api/v1/login", {
-            method: "POST",
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify(formData)
-        }).then(response => {
-            if (response.ok) {
-                response.text().then(text => {
-                    console.log(text) // todo set global variable
-                });
-                window.location.reload();
-            }
-        }).catch(error => {
-            alert(error)
-            console.error(error);
-        });
-    };
-
     return (
-        <div className={"login--form"} onSubmit={handleLoginSubmit}>
+        <div className={"login--form"} onSubmit={handleLoginSubmit(formData)}>
             <form>
                 <div className="input--container">
                     <label>Email </label>
